Ft_printf ALL

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ogoman <ogoman@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 15:58:57 by ogoman            #+#    #+#             */
-/*   Updated: 2023/11/15 09:49:33 by ogoman           ###   ########.fr       */
+/*   Updated: 2023/11/16 11:17:54 by ogoman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,9 +31,9 @@ void	ft_format(va_list ap, char *str, size_t *counter)
 	else if (*str == 'x' || *str == 'X')
 	{
 		if (*str == 'x')
-			ft_puthex_ft(va_arg(ap, uintmax_t), counter, HEX_LOW);
+			ft_puthex_ft(va_arg(ap, unsigned int), counter, HEX_LOW);
 		else
-			ft_puthex_ft(va_arg(ap, uintmax_t), counter, HEX_UPP);
+			ft_puthex_ft(va_arg(ap, unsigned int), counter, HEX_UPP);
 	}
 }
