+++ ft_puthex_ft.c
@@ -6,13 +6,13 @@
 /*   By: ogoman <ogoman@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 09:38:46 by ogoman            #+#    #+#             */
-/*   Updated: 2023/11/15 09:57:58 by ogoman           ###   ########.fr       */
+/*   Updated: 2023/11/16 11:18:20 by ogoman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_puthex_ft(uintmax_t num, size_t *counter, int letter_case)
+void	ft_puthex_ft(unsigned int num, size_t *counter, int letter_case)
 {
 	if (num >= 16)
 		ft_puthex_ft(num / 16, counter, letter_case);
